Switch from qemu-debootstrap to debootstrap.
Support for the former was removed in vmdb2 0.27, and unstable/testing
have 0.28 now. The switch is known to work in bookworm as well (0.26).

It's supposed to be transparent, but due to different options in both
plugins, one has to specify require_empty_target for the debootstrap
plugin (which wasn't needed for qemu-debootstrap).